+++ ton-kotlin-asm/src/commonMain/kotlin/org/ton/asm/contstack/SETCONTARGS_N.kt
@@ -1,10 +1,11 @@
 package org.ton.asm.contstack
 
 import org.ton.asm.AsmInstruction
+import org.ton.bigint.toUByte
 import org.ton.cell.CellBuilder
 import org.ton.cell.CellSlice
+import org.ton.cell.storeUInt
 import org.ton.tlb.TlbConstructor
-import org.ton.bigint.*
 import org.ton.tlb.providers.TlbConstructorProvider
 
 data class SETCONTARGS_N(
@@ -25,7 +26,7 @@ private object SETCONTARGS_NTlbConstructor : TlbConstructor<SETCONTARGS_N>(
 ) {
     override fun storeTlb(cellBuilder: CellBuilder, value: SETCONTARGS_N) {
         cellBuilder.storeUInt(value.r, 4)
-        cellBuilder.storeUIntLeq(value.n, 14)
+        cellBuilder.storeUIntLeq(value.n.toInt(), 14)
     }
 
     override fun loadTlb(cellSlice: CellSlice): SETCONTARGS_N {
